Black pre-commit.
